+++ src/eslint/base.cjs
@@ -16,6 +16,9 @@ module.exports = {
     parserOptions: {
         ecmaVersion: "latest",
         sourceType: "module",
+        ecmaFeatures: {
+            jsx: true,
+        },
     },
     plugins: ["@typescript-eslint", "prettier", "import"],
     overrides: [